[REFACTOR] 바뀐 폴더 경로 업로드
