+++ backendPlan/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'neroplan'
+rootProject.name = "backendPlan"
