+++ backendPlan/src/main/java/com/neroplan/backendPlan/BackendPlanApplication.java
@@ -1,13 +1,13 @@
-package com.neroplan.neroplan;
+package com.neroplan.backendPlan;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class NeroplanApplication {
+public class BackendPlanApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(NeroplanApplication.class, args);
+		SpringApplication.run(BackendPlanApplication.class, args);
 	}
 
 }
