+++ backendPlan/src/main/java/com/neroplan/backendPlan/domain/plan/dto/UpdatePlanRequestDto.java
@@ -1,4 +1,4 @@
-package com.neroplan.neroplan.domain.plan.dto;
+package com.neroplan.backendPlan.domain.plan.dto;
 
 import lombok.*;
